Move RBAC auth handler to authorization module
