+++ libsplinter/src/rest_api/auth/authorization/mod.rs
@@ -18,3 +18,4 @@
 pub mod allow_keys;
 #[cfg(feature = "authorization-handler-maintenance")]
 pub mod maintenance;
+pub mod rbac;
