+++ libsplinter/src/rest_api/auth/authorization/rbac/store/diesel/operations/add_role.rs
@@ -12,12 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use diesel::{
-    dsl::{delete, insert_into, update},
-    prelude::*,
-};
+use diesel::{dsl::insert_into, prelude::*};
 
-use crate::rest_api::auth::rbac::store::{
+use crate::rest_api::auth::authorization::rbac::store::{
     diesel::{
         models::{RoleModel, RolePermissionModel},
         schema::{role_permissions, roles},
@@ -27,24 +24,19 @@ use crate::rest_api::auth::rbac::store::{
 
 use super::RoleBasedAuthorizationStoreOperations;
 
-pub trait RoleBasedAuthorizationStoreUpdateRole {
-    fn update_role(&self, role: Role) -> Result<(), RoleBasedAuthorizationStoreError>;
+pub trait RoleBasedAuthorizationStoreAddRole {
+    fn add_role(&self, role: Role) -> Result<(), RoleBasedAuthorizationStoreError>;
 }
 
 #[cfg(feature = "sqlite")]
-impl<'a> RoleBasedAuthorizationStoreUpdateRole
+impl<'a> RoleBasedAuthorizationStoreAddRole
     for RoleBasedAuthorizationStoreOperations<'a, diesel::sqlite::SqliteConnection>
 {
-    fn update_role(&self, role: Role) -> Result<(), RoleBasedAuthorizationStoreError> {
+    fn add_role(&self, role: Role) -> Result<(), RoleBasedAuthorizationStoreError> {
         let (role, permissions): (RoleModel, Vec<RolePermissionModel>) = role.into();
 
         self.conn.transaction::<_, _, _>(|| {
-            delete(role_permissions::table.filter(role_permissions::role_id.eq(&role.id)))
-                .execute(self.conn)?;
-
-            update(roles::table.find(&role.id))
-                .set(roles::display_name.eq(&role.display_name))
-                .execute(self.conn)?;
+            insert_into(roles::table).values(role).execute(self.conn)?;
 
             insert_into(role_permissions::table)
                 .values(permissions)
@@ -56,19 +48,13 @@ impl<'a> RoleBasedAuthorizationStoreUpdateRole
 }
 
 #[cfg(feature = "role-based-authorization-store-postgres")]
-impl<'a> RoleBasedAuthorizationStoreUpdateRole
+impl<'a> RoleBasedAuthorizationStoreAddRole
     for RoleBasedAuthorizationStoreOperations<'a, diesel::pg::PgConnection>
 {
-    fn update_role(&self, role: Role) -> Result<(), RoleBasedAuthorizationStoreError> {
+    fn add_role(&self, role: Role) -> Result<(), RoleBasedAuthorizationStoreError> {
         let (role, permissions): (RoleModel, Vec<RolePermissionModel>) = role.into();
-
         self.conn.transaction::<_, _, _>(|| {
-            delete(role_permissions::table.filter(role_permissions::role_id.eq(&role.id)))
-                .execute(self.conn)?;
-
-            update(roles::table.find(&role.id))
-                .set(roles::display_name.eq(&role.display_name))
-                .execute(self.conn)?;
+            insert_into(roles::table).values(role).execute(self.conn)?;
 
             insert_into(role_permissions::table)
                 .values(permissions)
