+++ libsplinter/src/rest_api/auth/authorization/rbac/store/diesel/operations/get_assigned_roles.rs
@@ -1,4 +1,4 @@
-// Copyright 2018-2020 Cargill Incorporated
+// Copyright 2018-2021 Cargill Incorporated
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -16,33 +16,53 @@ use std::convert::TryInto;
 
 use diesel::prelude::*;
 
-use crate::rest_api::auth::rbac::store::{
+use crate::rest_api::auth::authorization::rbac::store::{
     diesel::{
-        models::{RoleModel, RolePermissionModel},
-        schema::roles,
+        models::{AssignmentModel, IdentityModel, RoleModel, RolePermissionModel},
+        schema::{identities, roles},
     },
-    Role, RoleBasedAuthorizationStoreError,
+    Identity, Role, RoleBasedAuthorizationStoreError,
 };
 
 use super::RoleBasedAuthorizationStoreOperations;
 
-pub trait RoleBasedAuthorizationStoreListRoles {
-    fn list_roles(
+pub trait RoleBasedAuthorizationStoreGetAssignedRoles {
+    fn get_assigned_roles(
         &self,
+        identity: &Identity,
     ) -> Result<Box<dyn ExactSizeIterator<Item = Role>>, RoleBasedAuthorizationStoreError>;
 }
 
-impl<'a, C> RoleBasedAuthorizationStoreListRoles for RoleBasedAuthorizationStoreOperations<'a, C>
+impl<'a, C> RoleBasedAuthorizationStoreGetAssignedRoles
+    for RoleBasedAuthorizationStoreOperations<'a, C>
 where
     C: diesel::Connection,
     String: diesel::deserialize::FromSql<diesel::sql_types::Text, C::Backend>,
+    i16: diesel::deserialize::FromSql<diesel::sql_types::SmallInt, C::Backend>,
 {
-    fn list_roles(
+    fn get_assigned_roles(
         &self,
+        identity: &Identity,
     ) -> Result<Box<dyn ExactSizeIterator<Item = Role>>, RoleBasedAuthorizationStoreError> {
+        let search_identity = match identity {
+            Identity::Key(ref key) => key,
+            Identity::User(ref user_id) => user_id,
+        };
         self.conn
             .transaction::<Box<dyn ExactSizeIterator<Item = Role>>, _, _>(|| {
-                let roles = roles::table.load::<RoleModel>(self.conn)?;
+                let identities = identities::table
+                    .filter(identities::identity.eq(search_identity))
+                    .load::<IdentityModel>(self.conn)?;
+
+                let role_ids = AssignmentModel::belonging_to(&identities)
+                    .load::<AssignmentModel>(self.conn)?
+                    .into_iter()
+                    .map(|assignment| assignment.role_id)
+                    .collect::<Vec<_>>();
+
+                let roles = roles::table
+                    .filter(roles::id.eq_any(role_ids))
+                    .load::<RoleModel>(self.conn)?;
 
                 let perms = RolePermissionModel::belonging_to(&roles)
                     .load::<RolePermissionModel>(self.conn)?
