+++ libsplinter/src/rest_api/auth/authorization/rbac/store/diesel/operations/list_assignments.rs
@@ -16,7 +16,7 @@ use std::convert::TryInto;
 
 use diesel::prelude::*;
 
-use crate::rest_api::auth::rbac::store::{
+use crate::rest_api::auth::authorization::rbac::store::{
     diesel::{
         models::{AssignmentModel, IdentityModel},
         schema::identities,
