+++ libsplinter/src/rest_api/auth/authorization/rbac/store/error.rs
