+++ splinterd/src/daemon.rs
@@ -63,7 +63,7 @@ use splinter::rest_api::auth::authorization::allow_keys::AllowKeysAuthorizationH
 #[cfg(feature = "authorization-handler-maintenance")]
 use splinter::rest_api::auth::authorization::maintenance::MaintenanceModeAuthorizationHandler;
 #[cfg(feature = "authorization-handler-rbac")]
-use splinter::rest_api::auth::rbac::{
+use splinter::rest_api::auth::authorization::rbac::{
     rest_api::RoleBasedAuthorizationResourceProvider, RoleBasedAuthorizationHandler,
 };
 #[cfg(feature = "authorization")]
